moved a class to dedicated module
moved "Grid" class from "main.py" to dedicated module ("grid.py")
